requirements.txt
# Container Depth Estimation — Requirements
# Python 3.10+ recommended

# ── Core frameworks ──────────────────────────────────────────────────────────
streamlit>=1.32.0

# ── PyTorch (install separately for CUDA if needed) ──────────────────────────
# CPU: pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
# CUDA 12.1: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
torch>=2.1.0
torchvision>=0.16.0

# ── YOLO (ultralytics) ───────────────────────────────────────────────────────
ultralytics>=8.2.0

# ── Depth Anything V2 (via HuggingFace transformers) ─────────────────────────
transformers>=4.40.0
huggingface-hub>=0.22.0

# ── SAM2 ─────────────────────────────────────────────────────────────────────
# Install with:
#   pip install git+https://github.com/facebookresearch/segment-anything-2.git
# OR if available on PyPI:
# samv2>=1.0.0

# ── Computer Vision ───────────────────────────────────────────────────────────
opencv-python>=4.9.0
Pillow>=10.0.0

# ── Numerical / ML ───────────────────────────────────────────────────────────
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.4.0

# ── Visualization ─────────────────────────────────────────────────────────────
plotly>=5.20.0

# ── Misc ─────────────────────────────────────────────────────────────────────
tqdm>=4.66.0
